+++ tests/test_requests_mock_flask.py
@@ -1221,17 +1221,26 @@ def test_overlapping_routes_multiple_requests(mock_ctx: _MockCtxType) -> None:
     assert mock_response_base_2.text == expected_base_data.decode()
 
 
-def test_unknown_mock_type() -> None:
+def test_unknown_mock_module() -> None:
     """
-    When an unknown mock type is passed in, an error is raised.
+    When an unknown mock module is passed in, an error is raised.
     """
+    app = Flask(import_name=__name__, static_folder=None)
+
+    @app.route(rule="/")
+    def _() -> str:
+        """
+        Return a simple message.
+        """
+        return ""  # pragma: no cover
+
     expected_error = (
         "Expected a HTTPretty, ``requests_mock``, or ``responses`` object, "
-        "got <class 'object'>."
+        "got module 'json'."
     )
     with pytest.raises(expected_exception=TypeError, match=expected_error):
         add_flask_app_to_mock(
-            mock_obj=object(),
-            flask_app=Flask(import_name=__name__, static_folder=None),
+            mock_obj=json,
+            flask_app=app,
             base_url="http://www.example.com",
         )